upd: LTX-2.5 training support (trainer adaptation, stability, UI model streams)
- re-applied partial gradient checkpointing ratio and fp32 RoPE on main's core
- trainer adapted to the new core: split-pack support (video/audio VAE paths,
  embedding_weight_paths, scale factors from checkpoint metadata), batched
  encode API, tiling/perturbation migrations; obsolete inference.py removed
- stability: fp32 elementwise loss residuals, non-finite loss guard
- preprocessing scripts accept --video-vae-path / --audio-vae-path;
  process_captions adopted from main (sharded multi-GPU, atomic writes)
- ui: model selection (LTX-2.3 / LTX-2.5) with per-stream settings,
  per-stream precomputed roots (.precomputed/<stream>/) with provenance.json,
  exact (bucket, stream) dataset merges, stream-aware training runs
- local additions: merge_lora_to_base script, run.sh node bootstrap

+++ packages/ltx-core/src/ltx_core/model/transformer/transformer_args.py
@@ -250,7 +250,8 @@ class TransformerArgsPreprocessor:
         pe = precompute_freqs_cis(
             positions,
             dim=inner_dim,
-            out_dtype=x_dtype,
+            # float32 (not x_dtype): positional precision at long sequence lengths.
+            out_dtype=torch.float32,
             theta=self.positional_embedding_theta,
             max_pos=max_pos,
             use_middle_indices_grid=use_middle_indices_grid,
